Fix SEH ABI issues.
MinGW provides something that is almost the same as the Itanium ABI, but
with subtle differences.  Fix the structure layouts to account for this.

Disable the C++ exception structure layout check with MinGW - it is
currently broken and causes *all* exceptions to break.  Disabling it
just causes C++ exceptions to crash the program, rather than all
exceptions.